requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
celery
redis
langgraph
langsmith
langchain-core
langchain-openai
langchain-chroma
langchain-community
langchain-classic
langchain-text-splitters
chromadb
tavily-python
pymupdf
pytesseract
Pillow
